# docs/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "LanGigaCards API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Language/{code}": {
      "get": {
        "tags": [
          "Language"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/User/languages/{languageCode}": {
      "get": {
        "tags": [
          "UserLanguages"
        ],
        "parameters": [
          {
            "name": "languageCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LanguageProfileDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LanguageProfileDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LanguageProfileDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/User/languages/switch": {
      "put": {
        "tags": [
          "UserLanguages"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchTargetLanguageDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchTargetLanguageDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchTargetLanguageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LanguageProfileDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LanguageProfileDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LanguageProfileDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/User/languages/{languageCode}/setup": {
      "put": {
        "tags": [
          "UserLanguages"
        ],
        "parameters": [
          {
            "name": "languageCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteLanguageSetupDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteLanguageSetupDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteLanguageSetupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LanguageProfileDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LanguageProfileDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LanguageProfileDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CompleteLanguageSetupDto": {
        "required": [
          "proficiencyLevel"
        ],
        "type": "object",
        "properties": {
          "proficiencyLevel": {
            "minLength": 1,
            "pattern": "^(Just Starting|Beginner|Intermediate|Advanced|Fluent)$",
            "type": "string"
          },
          "difficultyMode": {
            "pattern": "^$|^(A1|A2|B1|B1\\+|B2|C1|C2)$",
            "type": "string",
            "nullable": true
          },
          "categoryIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "LanguageProfileDto": {
        "type": "object",
        "properties": {
          "languageCode": {
            "type": "string",
            "nullable": true
          },
          "languageName": {
            "type": "string",
            "nullable": true
          },
          "proficiencyLevel": {
            "type": "string",
            "nullable": true
          },
          "difficultyMode": {
            "type": "string",
            "nullable": true
          },
          "isSetupCompleted": {
            "type": "boolean"
          },
          "categoryIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "currentStreak": {
            "type": "integer",
            "format": "int32"
          },
          "longestStreak": {
            "type": "integer",
            "format": "int32"
          },
          "totalXp": {
            "type": "integer",
            "format": "int32"
          },
          "level": {
            "type": "integer",
            "format": "int32"
          },
          "lastStudiedDeckId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "lastStudiedWordId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "lastStudiedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "SwitchTargetLanguageDto": {
        "required": [
          "languageCode"
        ],
        "type": "object",
        "properties": {
          "languageCode": {
            "maxLength": 10,
            "minLength": 1,
            "type": "string"
          },
          "languageName": {
            "maxLength": 100,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Token değerinizi girin.",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}